+++ kosogor/src/main/kotlin/tanvd/kosogor/utils/Accessors.kt
@@ -2,12 +2,15 @@
 
 package tanvd.kosogor.utils
 
+import com.gradle.publish.PluginBundleExtension
+import com.jfrog.bintray.gradle.BintrayExtension
 import org.gradle.api.Project
 import org.gradle.api.internal.HasConvention
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.tasks.SourceSetContainer
 import org.gradle.kotlin.dsl.getByName
 import org.gradle.kotlin.dsl.getPluginByName
+import org.gradle.plugin.devel.GradlePluginDevelopmentExtension
 import org.jfrog.gradle.plugin.artifactory.dsl.ArtifactoryPluginConvention
 
 //Generated accessors to use in plugin
@@ -24,6 +27,11 @@ val Project._sourceSets: SourceSetContainer
 
 fun Project._publishing(configure: PublishingExtension.() -> Unit) = extensions.configure("publishing", configure)
 
-fun org.gradle.api.Project._bintray(configure: com.jfrog.bintray.gradle.BintrayExtension.() -> Unit): Unit =
+fun Project._bintray(configure: BintrayExtension.() -> Unit): Unit =
         (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("bintray", configure)
+
+fun Project._pluginBundle(configure: PluginBundleExtension.() -> Unit) = (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("pluginBundle", configure)
+
+fun Project._gradlePlugin(configure: GradlePluginDevelopmentExtension.() -> Unit): Unit =
+        (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("gradlePlugin", configure)