+++ kosogor/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+# 1.0.1 - 2019-03-7
+### Added
+- PublishPlugin with support of Gradle plugins portal
+- Fix issue with unneeded delete of build dir
+
 # 1.0.0 - 2019-02-19
 ### Added
 - PublishJar with support of Artifactory and Bintray